feat: add melee range modifier as boon-bane (#1522)
This feature adds the new ability to have the melee range modifier on weapons specify a roll type (e.g. boon/bane or advantage/disadvantage) in addition to a number.  The roll type entered much match one of the settings, "What you want to call rolls with advantage (3d6kh2)" or "What you want to call rolls with disadvantage (3d6k/2)".  The update does include a migration for existing melee range modifier values, but you might need to check them.  Thanks to @ZEE for the suggestion.

This update also includes a fix for the dodge / parry modifier (now negative) and an ES localization update from @ForjaSalvaje